Refactor actions insto their own files
