+++ packages/search-ui/src/actions/index.js
@@ -1,0 +1,10 @@
+export { default as addFilter } from "./addFilter";
+export { default as clearFilters } from "./clearFilters";
+export { default as removeFilter } from "./removeFilter";
+export { default as reset } from "./reset";
+export { default as setCurrent } from "./setCurrent";
+export { default as setFilter } from "./setFilter";
+export { default as setResultsPerPage } from "./setResultsPerPage";
+export { default as setSearchTerm } from "./setSearchTerm";
+export { default as setSort } from "./setSort";
+export { default as trackClickThrough } from "./trackClickThrough";
