+++ packages/search-ui/src/actions/reset.js
@@ -1,0 +1,7 @@
+/**
+ * Reset search experience to initial state
+ *
+ */
+export default function reset() {
+  this._setState(this.startingState);
+}
