+++ packages/search-ui/src/actions/setCurrent.js
@@ -1,0 +1,12 @@
+/**
+ * Set the current page
+ *
+ * Will trigger new search
+ *
+ * @param current Integer
+ */
+export default function setCurrent(current) {
+  this._updateSearchResults({
+    current
+  });
+}
