+++ packages/search-ui/src/actions/setSort.js
@@ -1,0 +1,15 @@
+/**
+ * Set the current sort
+ *
+ * Will trigger new search
+ *
+ * @param sortField String
+ * @param sortDirection String ["asc"|"desc"]
+ */
+export default function setSort(sortField, sortDirection) {
+  this._updateSearchResults({
+    current: 1,
+    sortDirection,
+    sortField
+  });
+}
